rename stable to composablestable
